requirementes.txt
django
djoser
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mssql-django